+++ src/test/java/com/rollingpaper/ggeujeogggeujeog/unit/notification/NotificationFeedServiceTest.java
@@ -1,10 +1,10 @@
-package com.rollingpaper.ggeujeogggeujeog.notification.application;
+package com.rollingpaper.ggeujeogggeujeog.unit.notification;
 
 import static com.rollingpaper.ggeujeogggeujeog.common.fixture.NotificationTestFixture.*;
 import static com.rollingpaper.ggeujeogggeujeog.common.fixture.UserTestFixture.*;
 import static org.mockito.BDDMockito.*;
 
-import java.util.Optional;
+import java.util.Arrays;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -14,6 +14,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import com.rollingpaper.ggeujeogggeujeog.notification.application.NotificationFeedService;
 import com.rollingpaper.ggeujeogggeujeog.notification.exception.NoSuchNotification;
 import com.rollingpaper.ggeujeogggeujeog.notification.infrastructure.NotificationMapper;
 
@@ -40,22 +41,22 @@ class NotificationFeedServiceTest {
 	@DisplayName("알림을 id를 통해 조회한다.")
 	void findNotification() {
 		//given
-		given(notificationMapper.findById(anyLong())).willReturn(
-			Optional.of(TestNotification.NOTIFICATION1)
+		given(notificationMapper.findAllByUserId(anyLong())).willReturn(
+			Arrays.asList(TestNotification.NOTIFICATION1, TestNotification.NOTIFICATION2)
 		);
 
 		//when
 		notificationFeedService.findById(TestNotification.NOTIFICATION1.getId());
 
 		//then
-		then(notificationMapper).should(times(1)).findById(anyLong());
+		then(notificationMapper).should(times(1)).findAllByUserId(anyLong());
 	}
 
 	@Test
 	@DisplayName("알림이 삭제할 때, 알림이 없다면 예외가 발생한다.")
 	void deleteNotificationWithException() {
 		//given
-		given(notificationMapper.findById(anyLong())).willThrow(NoSuchNotification.class);
+		given(notificationMapper.findAllByUserId(anyLong())).willThrow(NoSuchNotification.class);
 
 		//when
 		Assertions.assertThrows(NoSuchNotification.class,
@@ -69,15 +70,15 @@ class NotificationFeedServiceTest {
 	@DisplayName("현재 알림을 조회한 뒤, 삭제한다.")
 	void deleteNotification() {
 		//given
-		given(notificationMapper.findById(anyLong())).willReturn(
-			Optional.of(TestNotification.NOTIFICATION1)
+		given(notificationMapper.findAllByUserId(anyLong())).willReturn(
+			Arrays.asList(TestNotification.NOTIFICATION1, TestNotification.NOTIFICATION2)
 		);
 
 		//when
 		notificationFeedService.deleteNotification(TestNotification.NOTIFICATION1.getId());
 
 		//then
-		then(notificationMapper).should(times(1)).findById(anyLong());
+		then(notificationMapper).should(times(1)).findAllByUserId(anyLong());
 		then(notificationMapper).should(times(1)).update(any());
 	}
 }
